+++ R/setup.R
@@ -39,12 +39,6 @@ setup_flir <- function(path = ".") {
   }
 
   ### Files
-  invisible(
-    fs::dir_copy(
-      system.file("rules/builtin", package = "flir"),
-      fs::path(flir_dir, "rules/builtin")
-    )
-  )
   if (!fs::file_exists(file.path(flir_dir, "cache_file_state.rds"))) {
     saveRDS(NULL, file.path(flir_dir, "cache_file_state.rds"))
   }
@@ -53,51 +47,4 @@ setup_flir <- function(path = ".") {
     paste("  -", list_linters(), collapse = "\n")
   )
   writeLines(config_content, file.path(flir_dir, "config.yml"))
-}
-
-#' Update the `flir` setup
-#'
-#' @description
-#'
-#' When `flir` is updated, it can ship new built-in rules or update existing
-#' ones. `update_flir()` will automatically add those new rules to the
-#' `flir/rules/builtin` folder. Custom rules stored in `flir/rules/custom`
-#' are not affected.
-#'
-#' @inheritParams setup_flir
-#'
-#' @return Can add new files in the `flir/rules` folder, doesn't return anything.
-#' @export
-#'
-#' @examples
-#' \dontrun{
-#'   update_flir()
-#' }
-update_flir <- function(path = ".") {
-  flir_dir <- file.path(path, "flir")
-  built_in_rules <- list.files(
-    fs::path(flir_dir, "rules/builtin"),
-    pattern = "\\.yml$"
-  )
-  updated_built_in_rules <- system.file(
-    list.files("rules/builtin", pattern = "\\.yml$"),
-    package = "flir"
-  )
-  new_built_in_rules <- setdiff(updated_built_in_rules, built_in_rules)
-
-  # Copy everything so that rules that already exist can also be updated.
-  fs::file_copy(
-    updated_built_in_rules,
-    paste0(flir_dir, "/rules/builtin/", basename(updated_built_in_rules))
-  )
-
-  cli::cli_alert_success("Updated existing rules.")
-  if (length(new_built_in_rules) > 0) {
-    cli::cli_alert_success(
-      "Added {length(new_built_in_rules)} rule{?s}: {gsub('\\.yml$', '', basename(new_built_in_rules))}."
-    )
-    cli::cli_alert_info(
-      "Don't forget to add them in flir/config.yml to use them."
-    )
-  }
 }