+++ NEWS.md
@@ -1,4 +1,4 @@
-# flir (development version)
+# flir (0.4.0)
 
 ## Breaking changes